pull request changes refactoring
