+++ client/app/components/bibliothek/style.scss
@@ -1,5 +1,14 @@
-.contentText {
+.content-text {
     width: 100%;
     height: 100%;
     background-color: white;
+}
+.title-header {
+    color: #00b0de;
+    textAlign: center;
+    marginTop: 50px;
+}
+.paragraph-header {
+    text-align: left;
+    margin-top: 60px;
 }
