+++ client/app/components/book-item/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import style from './style.scss';
-import {Button} from 'react-bootstrap'
+import {Button} from 'react-bootstrap';
 
 export default class BookItem extends React.Component {
     render() {
@@ -10,10 +10,10 @@ export default class BookItem extends React.Component {
                     <img className={style.bookItemCover} src="./../../assets/images/Foxy-Joxy-plays-a-trick_bookdash_01.jpg"/>
                     <div className={style.buttonGroupWrapper}>
                         <div className={style.buttonWrapper}>
-                            <Button className={style.buttonBookItem}  >Lesen</Button>
+                            <Button className={style.buttonBookItem}>Lesen</Button>
                         </div>
                         <div>
-                            <Button className={style.buttonBookItem} >Vorlesen</Button>
+                            <Button className={style.buttonBookItem}>Vorlesen</Button>
                         </div>
                     </div>
                 </div>
