+++ client/app/components/header/index.js
@@ -1,7 +1,3 @@
-/**
- * Created by Maxi on 24.04.17.
- */
-
 import React from 'react';
 import style from './style.scss';
 import {Nav, NavItem} from 'react-bootstrap';
