+++ client/app/components/header/style.scss
@@ -2,30 +2,28 @@
     text-align: center;
     background-color: blue;
 }
-.logoWrapper{
+.logo-wrapper {
     float: left;
     margin: 0px;
     padding: 0px;
 }
-.logo{
+.logo {
     width: 80%;
     margin-top: 10px;
 }
 
-.navBarWrapper
-{
+.nav-bar-wrapper {
     float:right;
     margin-top: 25px;
 }
-.navItem:first-child
-{
+.nav-item:first-child {
     margin-right: 10px;
 }
-.navItem{
+.nav-item {
     background-color: #00b0de;
     border-radius: 5px;
 }
 
-.navItem a{
+.nav-item a {
     color: white;
 }
