+++ client/app/components/main-layout/style.scss
@@ -1,3 +1,3 @@
-.head{
+.head {
     background-color: white;
 }
