+++ client/app/main.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import MainLayout from './components/main-layout';
 //import HelloWorldComponent from './components/hello-world';
 import Header from './components/header';
-import BibliothekPage from './components/bibliothek';
+import Library from './components/bibliothek';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
@@ -22,7 +22,7 @@ ReactDOM.render(
     <Provider store={ store }>
         <Router>
             <MainLayout header={ header }>
-                <Route exact path="/" component={ BibliothekPage }/>
+                <Route exact path="/" component={ Library }/>
             </MainLayout>
         </Router>
